Transfered the reachability code to the APIHelper struct
