+++ CKL iOS Challenge/CKL iOS Challenge/Helpers/APIHelper.swift
@@ -17,7 +17,7 @@ struct APIPaths {
     static let articleURL: String = "\(APIPaths.rootUrl)/article"
 }
 
-class APIHelper: NSObject {
+struct APIHelper {
     
     static func getArticlesList(_ context: NSManagedObjectContext, _ completion: @escaping (EZCoreDataResult<[Article]>) -> Void) {
         Alamofire.request(APIPaths.articleURL).validate().responseJSON { (response) in
@@ -32,6 +32,27 @@ class APIHelper: NSObject {
             case .failure(let error):
                 print(error)
                 completion(EZCoreDataResult<[Article]>.failure(error: error))
+            }
+        }
+    }
+}
+
+
+struct AppNotifications {
+    static let PhoneIsOffline = Notification.Name("PhoneIsOffline")
+    static let PhoneIsOnline = Notification.Name("PhoneIsOnline")
+}
+
+extension APIHelper {
+    static let reachabilityManager = NetworkReachabilityManager()
+    
+    static func setupReachability() {
+        reachabilityManager?.startListening()
+        reachabilityManager?.listener = { _ in
+            if let isNetworkReachable = self.reachabilityManager?.isReachable, isNetworkReachable == true {
+                NotificationCenter.default.post(name: AppNotifications.PhoneIsOnline, object: nil)
+            } else {
+                NotificationCenter.default.post(name: AppNotifications.PhoneIsOffline, object: nil)
             }
         }
     }
